Add support for displaying json keymaps
